partialScheduleMupa: do not assume the given node is a band
Original implementation assumed that the given node is band and used a
CHECK macro to enforce that.  In practice, it may be useful to call
partialScheduleMupa to obtain the partial schedule including the
schedule of the current node if it is a band, and the prefix schedule
otherwise.  Replace the CHECK by an early return of the result of
prefixScheduleMupa.