insertCopiesUnder: make groupId argument mandatory
The default null value of the groupId in insertCopiesUnder was left over
during a refactoring in the prehistory.  The function checks if it is
not null early, without any side effects before the check.  Remove the
default value of the argument and the check.  The group id is now
mandatory, and it is ensured by a compile-time rather than by a runtime
check.